Multiple bug fixes
Updated testcases with more bugs.

+++ final/src/packages/math.go
@@ -1,8 +1,8 @@
 package main
 
-var Pi float32 = 22 / 7
+var Pi float64 = 22.0 / 7
 
-func _shift(x float32) float32 {
+func _shift(x float64) float64 {
 	for (x < -Pi) {
 		x += 2 * Pi
 	}
@@ -12,13 +12,13 @@ func _shift(x float32) float32 {
 	return x
 }
 
-func Sin(x float32) float32 {
+func Sin(x float64) float64 {
 	x = _shift(x)
 	sin := x - x*x*x/6 + x*x*x*x*x/120
 	return sin
 }
 
-func Cos(x float32) float32 {
+func Cos(x float64) float64 {
 	x = _shift(x)
 	cos := 1 - x*x/2 + x*x*x*x/24
 	return cos
